+++ src/engine/mod.rs
@@ -9,6 +9,3 @@ use bot::Bot;
 
 mod player_response;
 use player_response::PlayerResponse;
-
-mod game_manager;
-pub use game_manager::GameManager;
